backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-dotenv==1.0.1
pydantic==2.10.4
boto3==1.35.86
playwright==1.49.1
python-multipart==0.0.20
supabase==2.11.0
httpx==0.28.1
